Remove tratamento de visitas no resident_control.py. Deleta test.py. Renomeia os arquivos: resident_control -> resident_auth e notify -> notify_admin.
