+++ bot/main.py
@@ -5,11 +5,11 @@ Start program
 import logging
 import os
 from commands import *
-from resident_control import Auth, HandleEntryVisitor
+from resident_auth import Auth
 from register import Register
 from register_visitor import RegisterVisitor
 from feedback import Feedback
-from notify import NotifyAdmin
+from notify_admin import NotifyAdmin
 from notify_resident import NotifyResident
 from visit import Visit
 from register_admin import RegisterAdmin
@@ -107,19 +107,6 @@ if __name__ == '__main__':
             },
 
         fallbacks=[CommandHandler('cancelar', Visit.end)]
-        ))
-
-    # Resident control (manage apartment entries )
-    dp.add_handler(ConversationHandler(
-        entry_points=[CommandHandler('autorizar', Auth.index)],
-
-        states={
-            CPF_AUTH:[MessageHandler(Filters.text, Auth.cpf)],
-            VOICE_AUTH: [MessageHandler(Filters.voice, Auth.voice)],
-            HANDLE_VISITORS_PENDING: [MessageHandler(Filters.text, HandleEntryVisitor.index)]
-            },
-
-        fallbacks=[CommandHandler('cancelar', HandleEntryVisitor.end)]
         ))
 
     # Feedback
